pyproject.toml
[project]
name = "slide-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "slide-lye>=0.3.0",
    "slide-tyler>=1.3.0",
    "python-dotenv",
    "slide-narrator>=1.0.1",
]
